Fix `preprocess`
When preprocessing (a.k.a. compiling the circuit or generating proving and verifying keys) we have to sample random witnesses because we don't know the ones that the user is going to use.

Nevertheless, the ACIR circuit does know this. We should find a way of knowing this too because this won't let us sample random circuit variables (witnesses) because they're pre-set in the ACIR circuit.